requirements.txt
# Core API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Logging
structlog==24.4.0

# OpenTelemetry - Observability
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-semantic-conventions==0.48b0

# Dashboard
streamlit==1.39.0
plotly==5.24.1
pandas==2.2.3

# Environment Management
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Langchain compatibility (if needed from previous labs)
langchain>=0.3.7
langchain-core>=0.3.63
chromadb>=0.5.18

# Embeddings
sentence-transformers==2.2.2
httpx